+++ DICOM/IO/FileReference.cs
@@ -22,7 +22,7 @@ namespace Dicom.IO {
 			private set;
 		}
 
-		/// <summary>File will be deleted when object is Disposed.</summary>
+		/// <summary>File will be deleted when object is <c>Disposed</c>.</summary>
 		public bool IsTempFile {
 			get;
 			internal set;
@@ -41,6 +41,22 @@ namespace Dicom.IO {
 				TemporaryFileRemover.Delete(Name);
 			else if (File.Exists(Name))
 				File.Delete(Name);
+		}
+
+		/// <summary>
+		/// Moves file and updates internal reference.
+		/// 
+		/// Calling this method will also remove set the <see cref="IsTempFile"/> property to <c>False</c>.
+		/// </summary>
+		/// <param name="dstFileName"></param>
+		public void Move(string dstFileName, bool overwrite = false) {
+			// delete if overwriting; let File.Move thow IOException if not
+			if (File.Exists(dstFileName) && overwrite)
+				File.Delete(dstFileName);
+
+			File.Move(Name, dstFileName);
+			Name = Path.GetFullPath(dstFileName);
+			IsTempFile = false;
 		}
 
 		public byte[] GetByteRange(int offset, int count) {